\name{\%+\%}
\alias{\%+\%}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
String concatenation
}
\description{
Concatenate two strings
}
\usage{
x \%+\% y
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x, y}{
 One character or vector character
}
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{insertConcAddin}}
}
\examples{
"abc"  \%+\% "123"
1:10  \%+\% " -> " \%+\% 10:1
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
