\name{insertConcAddin}
\alias{insertConcAddin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Insert \%+\%
}
\description{
Inserts "\%+\%" into text. This is needed for the "Insert \%+\%" button in Add-in.
}
\usage{
insertConcAddin()
}
%- maybe also 'usage' for other objects documented here.
\details{
You can see how to use this feature \href{../doc/example.html}{here}.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
