\name{insertConcAddin}
\alias{insertConcAddin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Insert \%+\%
}
\description{
Inserts "\%+\%" into text. This is needed for the "Insert \%+\%" button in Addins.
}
\usage{
insertConcAddin()
}
%- maybe also 'usage' for other objects documented here.
\details{
You can see how to use this feature \href{../doc/example.html}{here}.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
